data-access/app/requirements.txt
kafka-python == 2.0.2
flask
shortuuid